0-print_dlistint

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -9,7 +9,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	if (!h)
 		return (n);
